LayoutTests/platform/chromium/fast/events/touch/gesture/touch-gesture-scroll-iframe-expected.txt


This tests gesture event scrolling in iframes. Red-green strip is scrolled on pass.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
wheel event 0+> [object HTMLDivElement]
PASS touchtarget.contentDocument.body.scrollTop is 140
PASS touchtarget.contentDocument.body.scrollLeft is 0
PASS wheelEventsOccurred is 1
second gesture
wheel event 0+> [object HTMLDivElement]
PASS touchtarget.contentDocument.body.scrollTop is 200
PASS touchtarget.contentDocument.body.scrollLeft is 0
PASS wheelEventsOccurred is 1
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
